Feat: Refactor code base for better distinction between hackerrank challenges and game, write methods to generate and display board
