+++ lib/hacker_rank_challenges/save_princess1.rb
@@ -1,4 +1,4 @@
-require 'validations_helper'
+require_relative 'validations_helper'
 
 def displayPathtoPrincess(n, grid)
   validateInteger(n)
