+++ lib/hacker_rank_challenges/validations_helper.rb
